Move some stuff around, try simple DP binding
